backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[argon2]==1.7.4
email-validator==2.1.0
pydantic[email]==2.5.0
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
Pillow==10.1.0
